🎨 Design : 베너 작업
- [x] ✅ Done : 베너 이미지 반응형
- [x] ✅ Done : 베너 svg neon 스타일 ani 작업
