+++ src/app/globals.css
@@ -9,21 +9,29 @@
 }
 
 :root {
+
+  --base-color: rgb(15, 15, 241);
+  --base-color-light: hsl(208, 69%, 61%);
+
+
+
+
   /* title */
   --title-height: 64px;
   --title-backgroud-color: transparent;
-  --title-color: rgb(15, 15, 241);
-  --title-color-light: hsl(208, 69%, 61%);
+  --title-color: var(--base-color);
+  --title-color-light: var(--base-color-light);
   --title-font-size: 32px;
   --title-font-family-logo: 'Sacramento', cursive;
   --title-font-family-nav: 'Sacramento', cursive;
 
+  --max-width: 1920px;
 
 
 
 
 
-  --max-width: 1100px;
+
   --border-radius: 12px;
   --font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
     'Roboto Mono', 'Oxygen Mono', 'Ubuntu Monospace', 'Source Code Pro',
@@ -63,6 +71,8 @@
   --card-rgb: 180, 185, 188;
   --card-border-rgb: 131, 134, 135;
 }
+
+
 
 @media (prefers-color-scheme: dark) {
   :root {
@@ -107,21 +117,6 @@ html,
 body {
   max-width: 100vw;
   overflow-x: hidden;
-}
-
-body {
-  color: rgb(var(--foreground-rgb));
-  background: linear-gradient(
-      to bottom,
-      transparent,
-      rgb(var(--background-end-rgb))
-    )
-    rgb(var(--background-start-rgb));
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
 }
 
 @media (prefers-color-scheme: dark) {
