+++ src/app/layout.tsx
@@ -1,3 +1,4 @@
+import Nav from '@/components/Nav'
 import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
@@ -5,8 +6,8 @@ import { Inter } from 'next/font/google'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Kyoulee Blog',
+  description: 'Blog',
 }
 
 export default function RootLayout(
@@ -15,10 +16,10 @@ export default function RootLayout(
 ) {
   return (
     <html lang="en">
-      <body className={inter.className}>
-        
-        {children}
-      </body>
+    <body className={inter.className}>
+      <Nav />
+      {children}
+    </body>
     </html>
   )
 }
