--- src/app/styles/title.module.css
@@ -1,193 +1,0 @@
-.nav {
-    position: absolute;
-    margin-top: 8px;
-    width: 100%;
-}
-
-.contanier {
-    margin: auto;
-    position: relative;
-    height: var(--title-height);
-    width: 90%;
-    background-color: var(--title-backgroud-color);
-    border-radius: 32px;
-    overflow: hidden;
-    animation: 5s neonsign_box infinite alternate;
-}
-
-.content {
-    display: flex;
-}
-
-/* profile */
-.profile {
-    width: var(--title-height);
-    height: var(--title-height);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.profile .profile_img {
-    width: calc(var(--title-height) - 24px);
-    height: calc(var(--title-height) - 24px);
-    border-radius: 100%;
-    border: 1px solid var(--title-color-light);
-    box-shadow: 0px 0px 1px var(--title-color),
-        0px 0px 2px var(--title-color),
-        0px 0px 4px var(--title-color);
-    overflow: hidden;
-
-}
-
-.profile .profile_img img {
-    width: 100%;
-}
-
-/* logo */
-.nav .logo {
-    position: absolute;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    height: 100%;
-    font-family: var(--title-font-family-logo);
-    font-size: var(--title-font-size);
-    animation: neonsign_text 3s infinite alternate;
-}
-
-.nav .logo .kstyle {
-    background:
-        linear-gradient(10deg, var(--title-color) 0%, var(--title-color-light) 70%, var(--title-color-light) 100%);
-    background-clip: text;
-    color: transparent;
-    -webkit-box-reflect: below -12px linear-gradient(transparent, rgba(1, 1, 1, 50%));
-}
-
-/* nav guide */
-.navigate {
-    margin-left: auto;
-    display: inline-block;
-    overflow: hidden;
-}
-
-.navigate ul {
-    position: relative;
-    height: 100%;
-    display: flex;
-    align-items: end;
-    margin-right: 24px;
-    padding-bottom: 8px;
-    color: transparent;
-}
-
-.navigate ul li {
-    list-style-type: none;
-    margin-left: 12px;
-    max-width: 64px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    font-size: 16px;
-    font-family: var(--title-font-family-nav);
-    text-shadow:
-    0 0 1px #aaa,
-    0 0 2px #333,
-    0 0 4px var(--title-color),
-    0 0 8px var(--title-color),
-    0 0 16px var(--title-color),
-    0 0 36px var(--title-color),
-    0 0 64px var(--title-color);
-    transition-duration: 3s;
-}
-
-.navigate ul li:hover {
-    text-shadow:
-    0 0 1px #fff,
-    0 0 2px #aaa,
-    0 0 4px var(--title-color),
-    0 0 8px var(--title-color),
-    0 0 16px var(--title-color),
-    0 0 16px var(--title-color),
-    0 0 36px var(--title-color),
-    0 0 36px var(--title-color),
-    0 0 64px var(--title-color);
-}
-
-
-
-/* animation */
-@keyframes neonsign_text {
-    100%, 50% {
-        text-shadow:
-            0 0 1px #fff,
-            0 0 2px #fff,
-            0 0 4px var(--title-color),
-            0 0 8px var(--title-color),
-            0 0 8px var(--title-color),
-            0 0 16px var(--title-color),
-            0 0 16px var(--title-color),
-            0 0 36px var(--title-color),
-            0 0 36px var(--title-color),
-            0 0 36px var(--title-color),
-            0 0 36px var(--title-color),
-            0 0 64px var(--title-color),
-            0 0 64px var(--title-color),
-            0 0 64px var(--title-color),
-            0 0 64px var(--title-color),
-            0 0 128px var(--title-color);
-    }
-
-    5%, 0% {
-        text-shadow:
-            0 0 1px #444,
-            0 0 2px #444,
-            0 0 16px var(--title-color),
-            0 0 32px var(--title-color);
-    }
-}
-
-@keyframes neonsign_box {
-    100%, 50%, 20%, 10% {
-        box-shadow:
-            0 0 1px #fff,
-            0 0 2px #fff,
-            0 0 4px var(--title-color),
-            0 0 4px var(--title-color) inset,
-            0 0 8px var(--title-color),
-            0 0 8px var(--title-color),
-            0 0 8px var(--title-color) inset,
-            0 0 8px var(--title-color) inset,
-            0 0 16px var(--title-color),
-            0 0 16px var(--title-color),
-            0 0 16px var(--title-color) inset,
-            0 0 36px var(--title-color),
-            0 0 36px var(--title-color) inset;
-            
-    }
-
-    15%, 5%, 0% {
-        box-shadow:
-            0 0 1px #444,
-            0 0 2px #444,
-            0 0 16px var(--title-color),
-            0 0 16px var(--title-color) inset;
-    }
-}
-
-
-/* media width */
-
-@media (max-width: 740px) {
-    .navigate ul li:last-child {
-        display: none;
-        margin-right: 0px;
-    }
-}
-
-@media (max-width: 640px) {
-    .navigate ul {
-        display: none;
-    }
-}
